backend/requirements.txt
# NoirVision backend - TwelveLabs pipeline
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
boto3>=1.34.0
sqlmodel>=0.0.14
python-multipart>=0.0.6
PyJWT[crypto]>=2.8.0
